Implement the tinkers chisel
The chisel serves as an example no parts tool.
For gameplay, it is intended to make it easier to mine stone early game, and to make it easier to build late game via the exchanging modifier.
The chisel has two unique modifiers for example modifiers: one requires a slot, and one has a prerequisite