Revamp blowpipe
Blowpipe kinda sucks as a bow and its debatably harder to get, so its new gimick is producing free bamboo arrows instead of using normal arrows. Ideal for early game when you don't have a lot of ammo options
As part of this, removed it from the bow tag, meaning it had to manually get the relevant recipes. In this case, the only ones missing that we wanted were blocking and multishot